Fix/improve deduping logic re. legalName field
Discovered that APHIS appears to occasionally bulk-change the
`legalName` for any given `customerNumber`. Because the `sort_key`
incorporated that `legalName` and the scripts merge cached results with
newer ones, the previous logic was retaining multiple copies of the same
inspection (i.e., for inspections for which the `legalName` changed in
the fresh/recent results, but not yet in our historical cache). The good
news is that the `legalName` seems to correspond directly to
`customerNumber`, so we can just use that.

Also adjusted `get_sort_key` to account for other aspects observed in
the data, e.g., that `customerNumber` is never blank.