Don't throw error on extra fields in results
The metadata returned by the APHIS portal recently added some new
fields: `zip`, `state`, `city`, and `certType`. This causes
`csv.DictWriter` to throw an error when writing `inspections.csv`,
because the fieldnames for that CSV are based on our cached results,
which did not have those fields. Although we can get the same data via
the records we already have, we might still want to add these four
columns in the future. That will take a bit of work to get right so, in
the meantime, this commit simply adds the `extrasaction="ignore"`
parameter to the `csv.DictWriter` instantiation.